Updated kerasTensorflowClassifier.py to import model architectures rather than have them saved in the file - should make 'plug and playing' of different model architectures a bit easier to track.
